requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
